# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TravelMap API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/bookings:
    post:
      tags: [Booking]
      summary: Create a booking and hold capacity for 10 minutes
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateBookingRequest' } } }
      responses:
        '201': { description: Booking hold created, content: { application/json: { schema: { $ref: '#/components/schemas/BookingResponse' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { description: Capacity unavailable or POI is not bookable }
  /api/v1/bookings/external:
    post:
      tags: [Booking]
      summary: Import a Geoapify POI when needed, then create a booking hold
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateExternalBookingRequest' } } }
      responses:
        '201': { description: External POI imported/reused and booking hold created, content: { application/json: { schema: { $ref: '#/components/schemas/BookingResponse' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { description: Capacity unavailable or imported POI is not bookable }
  /api/v1/bookings/{bookingId}:
    get:
      tags: [Booking]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: bookingId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200': { description: Owned booking, content: { application/json: { schema: { $ref: '#/components/schemas/BookingResponse' } } } }
        '404': { description: Booking not found in caller scope }
  /api/v1/bookings/{bookingId}/payment-sessions:
    post:
      tags: [Payment]
      summary: Create or return the idempotent mock payment session
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: bookingId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 8, maxLength: 120 } }
      responses:
        '201': { description: Payment session, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentSessionResponse' } } } }
  /api/v1/bookings/{bookingId}/cancel:
    patch:
      tags: [Booking]
      summary: Cancel a pending/confirmed booking; refund paid deposit when at least two hours early
      security: [{ bearerAuth: [] }]
      parameters: [{ name: bookingId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200': { description: Cancelled booking, content: { application/json: { schema: { $ref: '#/components/schemas/BookingResponse' } } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateBookingRequest:
      type: object
      required: [poiId, visitAt, partySize]
      properties:
        poiId: { type: string, format: uuid }
        visitAt: { type: string, format: date-time }
        partySize: { type: integer, minimum: 1, maximum: 1000 }
        notes: { type: string, maxLength: 1000 }
    CreateExternalBookingRequest:
      type: object
      required: [externalId, name, category, address, latitude, longitude, visitAt, partySize]
      properties:
        externalId: { type: string, maxLength: 160, description: Stable external Geoapify identifier from the client result }
        name: { type: string, maxLength: 120 }
        category: { type: string, maxLength: 120 }
        address: { type: string, maxLength: 500 }
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
        visitAt: { type: string, format: date-time }
        partySize: { type: integer, minimum: 1, maximum: 1000 }
        notes: { type: string, maxLength: 1000 }
    BookingResponse:
      type: object
      properties:
        bookingId: { type: string, format: uuid }
        poiId: { type: string, format: uuid }
        poiName: { type: string }
        visitAt: { type: string, format: date-time }
        slotEndAt: { type: string, format: date-time }
        partySize: { type: integer }
        status: { type: string, enum: [PENDING, CONFIRMED, COMPLETED, CANCELLED, EXPIRED] }
        holdExpiresAt: { type: string, format: date-time, nullable: true }
        depositAmount: { type: number, format: decimal }
        currency: { type: string, example: VND }
        paymentRequired: { type: boolean }
...
    PaymentSessionResponse:
      type: object
      properties:
        paymentId: { type: string, format: uuid }
        bookingId: { type: string, format: uuid }
        gateway: { type: string, example: MOCK }
        status: { type: string, enum: [CREATED, PAID, FAILED, REFUNDED] }
        amount: { type: number, format: decimal }
        currency: { type: string, example: VND }
        paymentUrl: { type: string, format: uri }
...
    ApiError:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        fields:
          type: object
          additionalProperties: { type: string }
        timestamp: { type: string, format: date-time }
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ApiError' }
...